Use bootstrap 5 for styling the template (#205)
Bootstrap looks a little better for a document-oriented view like
Irydium than github-markdown-css. Also, some pieces of MyST (e.g.
panels) are based on their styling, so it's a logical choice, at least
for this stage of the project.

We use a (little bit of) copied code from github markdown css to improve
the look of code blocks.