Add MyST "glue" directives
This replaces the previous method of using Svelte syntax for
in-document variable substitution.

This also allows us to remove our dependency on mdsvex. Mdsvex is
a great piece of software, but trying to support both MyST *and*
svelte syntax in the same document is pretty cumbersome and creates
a steeper learning curve.